Refactor visualization modules
