+++ tests/DM_dynamic/test_dm_dynamic_integration.py
@@ -15,7 +15,7 @@ from unittest.mock import patch
 
 # Importar módulos del pipeline
 from DRAFTS import config
-from DRAFTS.image_utils import save_detection_plot, _calculate_dynamic_dm_range
+from DRAFTS.visualization import save_detection_plot, _calculate_dynamic_dm_range
 from DRAFTS.dynamic_dm_range import get_dynamic_dm_range_for_candidate
 
 def test_dm_dynamic_integration():
